Combined Applicant and Partner validations into one.  Removed tests for partner validation
